HAN: Fix: 컴포넌트 경로 재설정
