+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Header from './components/common/Header';
-import Home from './pages/Home';
+import Home from './routes/Home';
 import Footer from './components/common/Footer';
 
 const App = () => {
