+++ src/pages/home/HomeContent.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import HomeMain from '../../components/home/HomeMain';
+import HomeCommunity from '../../components/home/HomeCommunity';
+import HomeNews from '../../components/home/HomeNews';
+import HomeShopping from '../../components/home/HomeShopping';
+import HomeIcons from '../../components/home/HomeIcons';
+
+
+const HomeContent = () => {
+    return (
+        <div>
+            <HomeMain />
+            <HomeIcons />
+            <HomeCommunity />
+            <HomeNews />
+            <HomeShopping />
+        </div>
+    );
+};
+
+export default HomeContent;
