+++ src/pages/oauth/SignUpForm.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import FileUploadComponent from '../shop/FileUploadComponent';
+import FileUploadComponent from '../../components/shop/FileUploadComponent';
 import { TextField, Button, Typography } from '@mui/material';
 import userEvent from '@testing-library/user-event';
 import { useNavigate } from 'react-router';
